refactor: service 레이어 정리
