+++ src/service/index.js
@@ -1,0 +1,7 @@
+export { request } from "./api.js";
+export { initRouter, push } from "./router.js";
+export {
+  setStorageItem,
+  getStorageItem,
+  removeStorageItem,
+} from "./storage.js";
